refactor Linux Computer Use window backends
