+++ computer-use-linux/src/windowing/backends/mod.rs
@@ -1,0 +1,4 @@
+pub mod cosmic;
+pub mod gnome;
+pub mod hyprland;
+pub mod kwin;
